pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.3",
    "d>=0.2.2",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "langchain>=1.1.0",
    "langchain-nvidia-ai-endpoints>=1.0.0",
    "langchain-openai>=1.1.0",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.7",
]

[tool.ruff]
line-length = 130

[tool.ruff.lint]
extend-select = ["E501"]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true
